+++ src/Objects/Entity.js
